commit_20211216_13:56

+++ src/hyde/driver/dataset/ground_network/ws/drv_data_ws_io.py
@@ -438,7 +438,8 @@ class DriverData:
             folder_name_anc_list = list_folder(folder_name_anc_main)
             for folder_name_anc_step in folder_name_anc_list:
                 if os.path.exists(folder_name_anc_step):
-                    os.rmdir(folder_name_anc_step)
+                    if not any(os.scandir(folder_name_anc_step)):
+                        os.rmdir(folder_name_anc_step)
 
     # -------------------------------------------------------------------------------------
